+++ app_links/example/windows/runner/win32_window.cpp
@@ -4,7 +4,7 @@
 
 #include "resource.h"
 
-#include "app_links_windows/app_links_windows_plugin.h"
+#include "app_links/app_links_plugin_c_api.h"
 
 namespace {
 
@@ -138,7 +138,7 @@ bool Win32Window::CreateAndShow(const std::wstring& title,
 bool Win32Window::SendAppLinkToInstance(const std::wstring& title) {
   // Find our exact window
   HWND hwnd = ::FindWindow(kWindowClassName, title.c_str());
-
+  
   if (hwnd) {
     // Dispatch new link to current window
     SendAppLink(hwnd);
@@ -146,20 +146,21 @@ bool Win32Window::SendAppLinkToInstance(const std::wstring& title) {
     // (Optional) Restore our window to front in same state
     WINDOWPLACEMENT place = { sizeof(WINDOWPLACEMENT) };
     GetWindowPlacement(hwnd, &place);
-    switch(place.showCmd) {
-      case SW_SHOWMAXIMIZED:
-          ShowWindow(hwnd, SW_SHOWMAXIMIZED);
-          break;
-      case SW_SHOWMINIMIZED:
-          ShowWindow(hwnd, SW_RESTORE);
-          break;
-      default:
-          ShowWindow(hwnd, SW_NORMAL);
-          break;
+
+    switch (place.showCmd) {
+    case SW_SHOWMAXIMIZED:
+        ShowWindow(hwnd, SW_SHOWMAXIMIZED);
+        break;
+    case SW_SHOWMINIMIZED:
+        ShowWindow(hwnd, SW_RESTORE);
+        break;
+    default:
+        ShowWindow(hwnd, SW_NORMAL);
+        break;
     }
+
     SetWindowPos(0, HWND_TOP, 0, 0, 0, 0, SWP_SHOWWINDOW | SWP_NOSIZE | SWP_NOMOVE);
     SetForegroundWindow(hwnd);
-    // END Restore
 
     // Window has been found, don't create another one.
     return true;