+++ app_links_macos/example/lib/main.dart
@@ -1,3 +1,0 @@
-void main() {
-  // See app_links main package example
-}