+++ app_links_macos/lib/app_links_macos.dart
@@ -1,3 +1,0 @@
-/// The app_links_platform_interface defaults to MethodChannelAppLinks
-/// as its instance, which is all the macOS implementation needs. This file
-/// is here to silence warnings when publishing to pub.