+++ app_links_windows/lib/app_links_windows.dart
@@ -1,3 +1,0 @@
-// The app_links_platform_interface defaults to MethodChannelAppLinks
-// as its instance, which is all the Windows implementation needs. This file
-// is here to silence warnings when publishing to pub.